+++ ExilenceNextApp/src/components/item-table/ItemTableContainer.tsx
@@ -1,18 +1,28 @@
-import { Box, Button, Grid, makeStyles, Theme } from '@material-ui/core';
-import WarningIcon from '@material-ui/icons/Warning';
+import {
+  Box,
+  Grid,
+  IconButton,
+  makeStyles,
+  Theme,
+  useTheme
+} from '@material-ui/core';
+import MoreHorizIcon from '@material-ui/icons/MoreHoriz';
 import { inject, observer } from 'mobx-react';
-import React, { ChangeEvent } from 'react';
+import React, { ChangeEvent, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { statusColors } from '../../assets/themes/exilence-theme';
+import {
+  statusColors,
+  primaryLighter
+} from '../../assets/themes/exilence-theme';
+import { ITableItem } from '../../interfaces/table-item.interface';
 import { AccountStore } from '../../store/accountStore';
 import { SignalrStore } from '../../store/signalrStore';
 import { UiStateStore } from '../../store/uiStateStore';
-import { exportData } from '../../utils/export.utils';
+import { mapPricedItemToTableItem } from '../../utils/item.utils';
 import ItemTableFilter from './item-table-filter/ItemTableFilter';
 import ItemTable, { Order } from './ItemTable';
-import { IPricedItem } from '../../interfaces/priced-item.interface';
-import { ITableItem } from '../../interfaces/table-item.interface';
-import { mapPricedItemToTableItem } from '../../utils/item.utils';
+import FilterListIcon from '@material-ui/icons/FilterList';
+import ItemTableMenuContainer from './item-table-menu/ItemTableMenuContainer';
 
 interface ItemTableContainerProps {
   uiStateStore?: UiStateStore;
@@ -32,6 +42,9 @@ const useStyles = makeStyles((theme: Theme) => ({
   placeholder: {
     display: 'flex',
     alignSelf: 'flex-end'
+  },
+  inlineIcon: {
+    color: primaryLighter
   },
   warning: {
     color: statusColors.warning
@@ -54,6 +67,7 @@ const ItemTableContainer: React.FC<ItemTableContainerProps> = ({
   const { activeGroup } = signalrStore!;
   const { t } = useTranslation();
   const classes = useStyles();
+  const theme = useTheme();
 
   let timer: NodeJS.Timeout | undefined = undefined;
 
@@ -91,6 +105,10 @@ const ItemTableContainer: React.FC<ItemTableContainerProps> = ({
     }
   };
 
+  const handleItemTableMenuOpen = (event: React.MouseEvent<HTMLElement>) => {
+    uiStateStore!.setItemTableMenuAnchor(event.currentTarget);
+  };
+
   return (
     <>
       <Box mb={itemTableFilterSpacing} className={classes.itemTableFilter}>
@@ -100,7 +118,7 @@ const ItemTableContainer: React.FC<ItemTableContainerProps> = ({
           justify="space-between"
           alignItems="center"
         >
-          <Grid item md={3}>
+          <Grid item md={5}>
             <ItemTableFilter
               array={getItems()}
               handleFilter={handleFilter}
@@ -108,14 +126,24 @@ const ItemTableContainer: React.FC<ItemTableContainerProps> = ({
             />
           </Grid>
           <Grid item className={classes.actionArea}>
-            <Button
-              color="primary"
-              variant="contained"
-              disabled={getItems().length === 0}
-              onClick={() => exportData(getItems())}
+            <IconButton
+              size="small"
+              className={classes.inlineIcon}
+              onClick={() =>
+                uiStateStore!.setShowItemTableFilter(
+                  !uiStateStore!.showItemTableFilter
+                )
+              }
             >
-              {t('label.net_worth_export')}
-            </Button>
+              <FilterListIcon />
+            </IconButton>
+            <IconButton
+              size="small"
+              className={classes.inlineIcon}
+              onClick={handleItemTableMenuOpen}
+            >
+              <MoreHorizIcon />
+            </IconButton>
           </Grid>
         </Grid>
       </Box>
@@ -130,6 +158,7 @@ const ItemTableContainer: React.FC<ItemTableContainerProps> = ({
           uiStateStore!.setItemTableOrderBy(col)
         }
       />
+      <ItemTableMenuContainer />
     </>
   );
 };